+++ src/main/java/com/innowise/repo/FeedbackRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.Feedback;
 import java.util.List;
 
 public interface FeedbackRepository {
-    Feedback findById(Long id);
+    Feedback findById(Integer id);
     List<Feedback> findAll();
 
     Feedback save(Feedback feedback);
     Feedback update(Feedback feedback);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
