+++ src/main/java/com/innowise/repo/TicketRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.Ticket;
 import java.util.List;
 
 public interface TicketRepository {
-    Ticket findById(Long id);
+    Ticket findById(Integer id);
     List<Ticket> findAll();
 
     Ticket save(Ticket ticket);
     Ticket update(Ticket ticket);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
