+++ src/main/java/com/innowise/repo/UserRoleRepository.java
@@ -5,11 +5,11 @@ import com.innowise.domain.UserRole;
 import java.util.List;
 
 public interface UserRoleRepository {
-    UserRole findById(Long id);
+    UserRole findById(Integer id);
     List<UserRole> findAll();
 
     UserRole save(UserRole userRole);
     UserRole update(UserRole userRole);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
