+++ src/main/java/com/innowise/repo/impl/HistoryRepositoryImpl.java
@@ -15,7 +15,7 @@ public class HistoryRepositoryImpl implements HistoryRepository {
     private Session session;
 
     @Override
-    public History findById(Long id) {
+    public History findById(Integer id) {
         return session.find(History.class, id);
     }
 
@@ -35,7 +35,7 @@ public class HistoryRepositoryImpl implements HistoryRepository {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Integer id) {
         Optional<History> historyToDelete = Optional.of(session.find(History.class, id));
         if (historyToDelete.isEmpty()) {
             //todo throw exception
